pytools/tw_institution_code/requirements.txt
aiobotocore==1.4.1
aiohttp==3.7.4.post0
aioitertools==0.8.0
async-timeout==3.0.1
attrs==21.2.0
botocore==1.20.106
chardet==4.0.0
fsspec==2021.8.1
idna==3.2
jmespath==0.10.0
multidict==5.1.0
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
s3fs==2021.8.1
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wrapt==1.12.1
yarl==1.6.3
